Move PenPreviewLabel to `widgets` folder
